Moving type enums & first metric class protocol
