+++ dbt_semantic_interfaces/objects/metric.py
@@ -2,7 +2,6 @@ from __future__ import annotations
 
 from typing import List, Optional
 
-from dbt_semantic_interfaces.enum_extension import ExtendedEnum
 from dbt_semantic_interfaces.errors import ParsingException
 from dbt_semantic_interfaces.objects.base import (
     HashableBaseModel,
@@ -12,21 +11,12 @@ from dbt_semantic_interfaces.objects.base import (
 )
 from dbt_semantic_interfaces.objects.filters.where_filter import WhereFilter
 from dbt_semantic_interfaces.objects.metadata import Metadata
-from dbt_semantic_interfaces.objects.time_granularity import (
+from dbt_semantic_interfaces.references import MeasureReference, MetricReference
+from dbt_semantic_interfaces.type_enums.metric_type import MetricType
+from dbt_semantic_interfaces.type_enums.time_granularity import (
     TimeGranularity,
     string_to_time_granularity,
 )
-from dbt_semantic_interfaces.references import MeasureReference, MetricReference
-
-
-class MetricType(ExtendedEnum):
-    """Currently supported metric types."""
-
-    MEASURE_PROXY = "measure_proxy"
-    RATIO = "ratio"
-    EXPR = "expr"
-    CUMULATIVE = "cumulative"
-    DERIVED = "derived"
 
 
 class MetricInputMeasure(PydanticCustomInputParser, HashableBaseModel):
