+++ tests/parsing/test_metric_parsing.py
@@ -7,9 +7,9 @@ from dbt_semantic_interfaces.objects.metric import (
     MetricTimeWindow,
     MetricType,
 )
-from dbt_semantic_interfaces.objects.time_granularity import TimeGranularity
 from dbt_semantic_interfaces.parsing.dir_to_model import parse_yaml_files_to_model
 from dbt_semantic_interfaces.parsing.objects import YamlConfigFile
+from dbt_semantic_interfaces.type_enums.time_granularity import TimeGranularity
 from dbt_semantic_interfaces.validations.validator_helpers import (
     ModelValidationException,
 )
